refactor: move createZip function to src/actions/export file
